requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
pydantic>=2.7.0
python-multipart>=0.0.9

pyserial>=3.5
websocket-client>=1.8.0
